Úpravy UI a pozvánek

+++ app/Livewire/Modals/Poll/Invitations.php
@@ -7,6 +7,8 @@ use App\Models\Poll;
 use App\Services\NotificationService;
 use Illuminate\Support\Str;
 use Livewire\Component;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
 
 class Invitations extends Component
 {
@@ -52,7 +54,7 @@ class Invitations extends Component
                 'email' => $invitation->email,
                 'status' => $invitation->status,
                 'key' => $invitation->key,
-                'sent_at' => $invitation->sent_at,
+                'sent_at' => Carbon::parse($invitation->sent_at)->diffForHumans(),
             ];
         }
     }
@@ -60,14 +62,16 @@ class Invitations extends Component
     // Metoda pro přidání pozvánky
     public function addInvitation()
     {
+        // Kontrola, zda je uživatel přihlášen
+        if (!Auth::check()) {
+            session()->flash('error', 'You must be logged in to send invitations.');
+            return;
+        }
 
         $this->validate();
 
-        foreach ($this->poll->invitations as $invitation) {
-            if ($invitation->email === $this->email) {
-                // Vypsat error
-                return;
-            }
+        if($this->checkIfCanBeSent() === false) {
+            return;
         }
 
         $invitation = Invitation::create([
@@ -75,6 +79,7 @@ class Invitations extends Component
             'email' => $this->email,
             'status' => 'pending',
             'key' => Str::random(40),
+            'sent_at' => now(),
         ]);
 
         // Odeslání pozvánky
@@ -91,8 +96,60 @@ class Invitations extends Component
     public function removeInvitation($id)
     {
         $invitation = Invitation::find($id);
-        $invitation->delete();
+
+        if ($invitation) {
+            $invitation->delete();
+            session()->flash('success', 'Invitation deleted successfully.');
+        } else {
+            session()->flash('error', 'Invitation not found.');
+        }
+
         $this->loadInvitations();
+    }
+
+    public function resendInvitation($id)
+    {
+        if($this->checkIfCanBeSent() === false) {
+            return;
+        }
+
+        $invitation = Invitation::find($id);
+
+
+        if ($invitation) {
+            $this->notificationService->sendInvitation($invitation->email, $this->poll, $invitation->key);
+            session()->flash('success', 'Invitation resent successfully.');
+        } else {
+            session()->flash('error', 'Invitation not found.');
+        }
+    }
+
+
+    private function checkIfCanBeSent(): bool
+    {
+
+
+        if(count($this->poll->invitations) >= 2) {
+            session()->flash('error', 'You can only send 2 invitations to this poll.');
+            return false;
+        }
+
+
+        $todayInvitations = $this->poll->invitations->where('sent_at', '>=', now()->subDay())->count();
+
+        if ($todayInvitations >= 10) {
+            session()->flash('error', 'You can only send 10 invitations per day.');
+            return false;
+        }
+
+        foreach ($this->poll->invitations as $invitation) {
+            if ($invitation->email === $this->email) {
+                session()->flash('error', 'Invitation was already sent to this email.');
+                return false;
+            }
+        }
+
+        return true;
     }
 
     public function render()
